Weapon 클래스 생성, ProjectileStorage생성
WeaponBase생성 Entity에서 공격을 담당하는 부분으로
Shot을 통해 투사체 발사
매개변수로 ProjecitleStorage의 투사체들 발사

ProjectileStorage 투사체 따로 보관하는 클래스. <Projectile, float>형태로 float은 TIme.time을 저장. Dictionary의 Value - Time.time을 WeaponBase에서 attackRate와 비교하는 역할

+++ Assets/Scripts/Weapon/Projectile/Straight.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Straight : Projectile
 {
-    public override void Setup(GameObject owner, int damage = 1, float speed = 1, int range = 1, float duration = 1, float attackRate = 1, float size = 1)
+    public override void Setup(GameObject owner, float damage = 1, float speed = 1, float range = 1, float duration = 1, float attackRate = 1, float size = 1)
     {
         base.Setup(owner, damage, speed, range, duration, attackRate, size);
     }
